requirements.txt
Flask==2.3.2
qrcode[pil]
google-cloud-storage==2.11.0
google-cloud-firestore==2.10.0
gunicorn==21.2.0
